fix: 로또 당첨 등수 Enum 비교시 오타로 인한 버그 수정

+++ src/main/java/lotto/domain/LottoService.java
@@ -30,7 +30,7 @@ public class LottoService {
     public double calculateProfit(LottoResult lottoResult, int numberOfLotto) {
 
         double gain = 0;
-        int buyMoney = numberOfLotto * 1000;
+        double buyMoney = numberOfLotto * 1000;
 
         gain += lottoResult.getFifthCount() * WinningRank.FIFTH.getPrice();
         gain += lottoResult.getFourthCount() * WinningRank.FOURTH.getPrice();
